+++ packages/flutter_adaptive_cards_fs/lib/flutter_adaptive_cards_fs.dart
@@ -6,12 +6,13 @@ library;
 
 export 'package:flutter_adaptive_cards_fs/src/action/action_handler.dart';
 export 'package:flutter_adaptive_cards_fs/src/adaptive_cards_canvas.dart';
-
+export 'package:flutter_adaptive_cards_fs/src/flutter_raw_adaptive_card.dart'
+    show RawAdaptiveCard, RawAdaptiveCardState;
 export 'package:flutter_adaptive_cards_fs/src/hostconfig/host_config.dart';
-
 export 'package:flutter_adaptive_cards_fs/src/models/action_invoke.dart';
 export 'package:flutter_adaptive_cards_fs/src/models/adaptive_card_update.dart';
 export 'package:flutter_adaptive_cards_fs/src/models/choice.dart';
+export 'package:flutter_adaptive_cards_fs/src/models/data_query.dart';
 export 'package:flutter_adaptive_cards_fs/src/models/fact.dart';
 export 'package:flutter_adaptive_cards_fs/src/models/media_source.dart';